Create filter list component with filter functionality
